ServingLLMWithFastAPI/requirements.txt
python-dotenv
pydantic
fastapi[standard]
openai
uvicorn